requirements.txt
fastapi
uvicorn
openai
langchain
requests
python-dotenv